qodana.yaml update
