--- .github/workflows/qodana.yml
@@ -1,18 +1,0 @@
-name: Code Quality
-
-on:
-  push:
-    branches:
-      - main
-      - wip
-  pull_request:
-    branches:
-      - main
-
-version: 1.0
-jobs:
-  qodana:
-    - name: 'Qodana Scan'
-      uses: JetBrains/qodana-action@main
-      env:
-        QODANA_TOKEN: ${{ secrets.QODANA_TOKEN }}
