+++ CoffeeSpace.Application/Attributes/Decorator.cs
@@ -1,6 +1,6 @@
 namespace CoffeeSpace.Application.Attributes;
 
-[AttributeUsage(validOn: AttributeTargets.Class, AllowMultiple = false, Inherited = false)]
+[AttributeUsage(validOn: AttributeTargets.Class, Inherited = false)]
 public sealed class Decorator : Attribute
 {
 }
