+++ CoffeeSpace.OrderingApi/Application/Messaging/Mediator/Commands/Buyers/UpdateBuyerCommand.cs
@@ -1,5 +1,4 @@
 ﻿using CoffeeSpace.Domain.Ordering.BuyerInfo;
-using CoffeeSpace.Domain.Ordering.CustomerInfo;
 using Mediator;
 
 namespace CoffeeSpace.OrderingApi.Application.Messaging.Mediator.Commands.Buyers;
