+++ CoffeeSpace.OrderingApi/Application/Repositories/Abstractions/IBuyerRepository.cs
@@ -1,5 +1,4 @@
 ﻿using CoffeeSpace.Domain.Ordering.BuyerInfo;
-using CoffeeSpace.Domain.Ordering.CustomerInfo;
 
 namespace CoffeeSpace.OrderingApi.Application.Repositories.Abstractions;
