+++ CoffeeSpace.OrderingApi/Persistence/Configurations/BuyerConfiguration.cs
@@ -1,9 +1,8 @@
 ﻿using CoffeeSpace.Domain.Ordering.BuyerInfo;
-using CoffeeSpace.Domain.Ordering.CustomerInfo;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Metadata.Builders;
 
-namespace CoffeeSpace.OrderingApi.Persistence.EntityTypeConfigurations;
+namespace CoffeeSpace.OrderingApi.Persistence.Configurations;
 
 public sealed class BuyerConfiguration : IEntityTypeConfiguration<Buyer>
 {
