+++ CoffeeSpace.OrderingApi/Persistence/Configurations/PaymentInfoConfiguration.cs
@@ -2,7 +2,7 @@ using CoffeeSpace.Domain.Ordering.BuyerInfo.CardInfo;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Metadata.Builders;
 
-namespace CoffeeSpace.OrderingApi.Persistence.EntityTypeConfigurations;
+namespace CoffeeSpace.OrderingApi.Persistence.Configurations;
 
 public sealed class PaymentInfoConfiguration : IEntityTypeConfiguration<PaymentInfo>
 {
