+++ CoffeeSpace.OrderingApi/Persistence/OrderingDbContext.cs
@@ -1,10 +1,9 @@
 using System.Reflection;
 using CoffeeSpace.Domain.Ordering.BuyerInfo;
 using CoffeeSpace.Domain.Ordering.BuyerInfo.CardInfo;
-using CoffeeSpace.Domain.Ordering.CustomerInfo;
 using CoffeeSpace.Domain.Ordering.Orders;
 using CoffeeSpace.OrderingApi.Persistence.Abstractions;
-using CoffeeSpace.OrderingApi.Persistence.EntityTypeConfigurations;
+using CoffeeSpace.OrderingApi.Persistence.Configurations;
 using MassTransit.EntityFrameworkCoreIntegration;
 using Microsoft.EntityFrameworkCore;
